backend/requirements.txt
fastapi==0.117.1
passlib==1.7.4
pydantic-settings==2.2.1
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.1.1
python_bcrypt==0.3.2
SQLAlchemy==2.0.43
uvicorn==0.30.6